requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
conda==4.3.16
distlib==0.3.4
filelock==3.4.2
idna==3.3
pipenv==2022.1.8
platformdirs==2.4.1
pycosat==0.6.3
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.8
virtualenv==20.13.1
virtualenv-clone==0.5.7
